Revert "Hinta kentät hyväksyy nytten desimaali arvot"
This reverts commit 9029b2582ae4e79d7a77d102329c66301b6a3804.

+++ src/kayttoliittymapakkaus/IlmoitusIkkuna.java
@@ -28,7 +28,7 @@ public class IlmoitusIkkuna extends Ikkuna{
                 arvot = syottopaneeli.getArvot();
         try {
             int id = Integer.parseInt(arvot[0]);
-            double hinta = Double.parseDouble(arvot[2]);
+            double hinta = Integer.parseInt(arvot[2]);
             int veneTilaus_id = Integer.parseInt(arvot[3]);
             
             rekisteri.lisaaTieto(new Ilmoitus(id, arvot[1], hinta, veneTilaus_id));
@@ -46,7 +46,7 @@ public class IlmoitusIkkuna extends Ikkuna{
 
             int id = Integer.parseInt(arvot[0]);
             int veneTilaus_id = Integer.parseInt(arvot[3]);
-            double hinta = Double.parseDouble(arvot[2]);
+            double hinta = Integer.parseInt(arvot[2]);
             rekisteri.lisaaTieto(new Ilmoitus(id, arvot[1], hinta, veneTilaus_id));
         paivitaValintaLista();
     }
